auth-service/requirements.txt
django>=4.2
djangorestframework>=3.14
PyJWT>=2.8
requests>=2.31
